Update divide into several files
