+++ src/errors/baseError.ts
@@ -1,0 +1,10 @@
+class FM3BaseError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'FM3BaseError';
+  }
+}
+
+export {
+  FM3BaseError,
+};
